Move PBS client into a dedicated pbs-client crate
- Extract the Proxmox Backup Server protocol client (config, token auth,
  TLS fingerprint pinning, h2 transport, DataBlob, FastCDC chunker,
  manifest, naming, reader, writer, rest) into a standalone `pbs-client`
  workspace crate, mirroring unrar-rs. The PBS-only dependencies (h2,
  hyper, hyper-util, http-body-util, tokio-rustls, fastcdc) move with it.
- The backup adapter stays in the application and depends on the crate.
- Drop the crate's unit-test modules.
- Align the pulled config with the panel contract: the snapshot time is
  `backup_created` (DateTime<Utc>), converted to a Unix timestamp.
