+++ application/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2024"
 
 [dependencies]
 unrar = { path = "../unrar-rs" }
+pbs-client = { path = "../pbs-client" }
 tracing = { version = "0.1.41", features = ["max_level_debug"] }
 tracing-subscriber = { version = "0.3.19", features = ["chrono"] }
 tracing-appender = "0.2.3"
@@ -39,11 +40,6 @@ sha2 = "0.11.0"
 ddup-bak = "0.9.0"
 infer = "0.19.0"
 rustls = { version = "0.23.27", features = ["aws-lc-rs"] }
-tokio-rustls = "0.26.4"
-hyper = { version = "1.10", features = ["client", "http1"] }
-hyper-util = { version = "0.1", features = ["tokio"] }
-http-body-util = "0.1"
-h2 = "0.4"
 bytes = "1"
 ignore = "0.4.23"
 sha1 = "0.11.0"
@@ -57,7 +53,6 @@ clap = { version = "4.5.37", features = ["derive", "string"] }
 dialoguer = "0.12.0"
 sysinfo = "0.39.0"
 crc32fast = "1.4.2"
-fastcdc = "3.2.1"
 cap-std = "4.0.0"
 rustix = "1.0.7"
 cidr = { version = "0.3.1", features = ["serde"] }
