+++ application/src/server/backup/mod.rs
@@ -10,7 +10,6 @@ use std::sync::{Arc, atomic::AtomicU64};
 
 pub mod adapters;
 pub mod manager;
-pub mod pbs;
 pub mod transfer;
 
 pub enum Backup {
