+++ pbs-client/src/datablob.rs
@@ -1,28 +1,19 @@
 use super::error::PbsError;
 use sha2::{Digest, Sha256};
 
-/// PBS DataBlob magic for unencrypted, uncompressed payloads.
 pub const UNCOMPRESSED_BLOB_MAGIC: [u8; 8] = [66, 171, 56, 7, 190, 131, 112, 161];
-/// PBS DataBlob magic for unencrypted, zstd-compressed payloads.
 pub const COMPRESSED_BLOB_MAGIC: [u8; 8] = [49, 185, 88, 66, 111, 182, 163, 127];
 
-/// Size of the unencrypted DataBlob header: `magic[8] || crc[4]`.
 const HEADER_SIZE: usize = 12;
 
-/// Upper bound on a decoded blob, guarding against decompression bombs.
 const MAX_DECODED_BLOB: usize = 256 * 1024 * 1024;
 
-/// An encoded PBS DataBlob plus the metadata PBS needs to register it.
 pub struct EncodedBlob {
-    /// The on-wire blob: `magic[8] || crc_le[4] || payload`.
     pub data: Vec<u8>,
-    /// SHA-256 of the **plaintext** — PBS's chunk digest / dedup key.
     pub digest: [u8; 32],
-    /// Plaintext length in bytes (PBS `size`).
     pub plaintext_size: u64,
 }
 
-/// SHA-256 of a byte slice.
 pub fn sha256(data: &[u8]) -> [u8; 32] {
     let digest = Sha256::digest(data);
     let mut out = [0u8; 32];
@@ -36,12 +27,6 @@ fn crc32(data: &[u8]) -> u32 {
     hasher.finalize()
 }
 
-/// Encodes plaintext into an unencrypted PBS DataBlob.
-///
-/// zstd compression is applied only when it actually shrinks the data, matching
-/// PBS's own behaviour (otherwise the uncompressed magic is used). The CRC32 is
-/// little-endian and covers the payload bytes only (everything after the
-/// 12-byte header).
 pub fn encode_blob(plaintext: &[u8]) -> EncodedBlob {
     let digest = sha256(plaintext);
 
@@ -67,7 +52,6 @@ pub fn encode_blob(plaintext: &[u8]) -> EncodedBlob {
     }
 }
 
-/// Decodes an unencrypted PBS DataBlob, verifying its CRC.
 pub fn decode_blob(raw: &[u8]) -> Result<Vec<u8>, PbsError> {
     let magic = raw
         .get(..8)
@@ -97,52 +81,5 @@ pub fn decode_blob(raw: &[u8]) -> Result<Vec<u8>, PbsError> {
         Err(PbsError::Decode(
             "unknown or encrypted blob magic (encryption is not supported)".into(),
         ))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn uncompressed_blob_has_exact_header_and_roundtrips() {
-        let input = b"hello pbs";
-        let blob = encode_blob(input);
-
-        // Small/incompressible input keeps the uncompressed magic.
-        assert_eq!(blob.data.get(..8), Some(UNCOMPRESSED_BLOB_MAGIC.as_slice()));
-        assert_eq!(blob.plaintext_size, input.len() as u64);
-        assert_eq!(blob.digest, sha256(input));
-
-        // CRC is little-endian over the payload only (bytes after the header).
-        let payload = blob.data.get(12..).expect("payload present");
-        let mut hasher = crc32fast::Hasher::new();
-        hasher.update(payload);
-        assert_eq!(
-            blob.data.get(8..12),
-            Some(hasher.finalize().to_le_bytes().as_slice())
-        );
-
-        assert_eq!(decode_blob(&blob.data).expect("decodes"), input);
-    }
-
-    #[test]
-    fn compressible_blob_uses_compressed_magic_and_roundtrips() {
-        let input = vec![0u8; 8192];
-        let blob = encode_blob(&input);
-
-        assert_eq!(blob.data.get(..8), Some(COMPRESSED_BLOB_MAGIC.as_slice()));
-        assert_eq!(blob.digest, sha256(&input));
-        assert_eq!(decode_blob(&blob.data).expect("decodes"), input);
-    }
-
-    #[test]
-    fn corrupt_crc_is_rejected() {
-        let mut blob = encode_blob(b"tamper me").data;
-        // Flip a payload byte without fixing the CRC.
-        if let Some(byte) = blob.get_mut(12) {
-            *byte ^= 0xff;
-        }
-        assert!(decode_blob(&blob).is_err());
     }
 }
