+++ pbs-client/src/h2.rs
@@ -1,31 +1,14 @@
-//! HTTP/2-over-upgrade transport for the PBS backup protocol.
-//!
-//! PBS's writer/reader protocols are HTTP/2 spoken on a connection that starts
-//! as an HTTP/1.1 request and is upgraded (`101 Switching Protocols`) to the
-//! `proxmox-backup-protocol-v1` protocol. We therefore:
-//!   1. open a TLS connection with the pinned-fingerprint config (no ALPN),
-//!   2. send the HTTP/1.1 `GET /api2/json/{backup|reader}` with the upgrade
-//!      headers and the token Authorization,
-//!   3. take the upgraded stream and run an `h2` client handshake on it,
-//!   4. issue inner requests (`dynamic_index`, `dynamic_chunk`, ...) over h2.
-//!
-//! Authorization is only sent on the initial upgrade request; inner h2 requests
-//! are unauthenticated, matching the PBS client.
-
 use super::{auth, config::PbsConfig, error::PbsError, naming, tls};
 use bytes::Bytes;
 use std::{future::poll_fn, sync::Arc};
 
-/// h2 flow-control window PBS uses: `(1 << 31) - 2`.
 const WINDOW_SIZE: u32 = (1 << 31) - 2;
-/// h2 max frame size PBS uses: 4 MiB.
 const MAX_FRAME_SIZE: u32 = 4 * 1024 * 1024;
 
 fn transport<E: std::fmt::Display>(err: E) -> PbsError {
     PbsError::Transport(err.to_string().into())
 }
 
-/// Parses `https://host:port` into `(host, port)`, defaulting to port 8007.
 fn parse_host_port(base_url: &str) -> Result<(String, u16), PbsError> {
     let without_scheme = base_url
         .strip_prefix("https://")
@@ -45,8 +28,6 @@ fn parse_host_port(base_url: &str) -> Result<(String, u16), PbsError> {
     }
 }
 
-/// Builds the snapshot query shared by the backup (writer) and reader protocols:
-/// `store`, `backup-type`, `backup-id`, `backup-time`, and optional `ns`.
 pub fn snapshot_query(config: &PbsConfig, backup_id: &str, backup_time: i64) -> String {
     let mut params: Vec<(&str, String)> = vec![
         ("store", config.datastore.to_string()),
@@ -62,7 +43,6 @@ pub fn snapshot_query(config: &PbsConfig, backup_id: &str, backup_time: i64) -> 
     encode_query(&params)
 }
 
-/// Builds a `key=value&...` query string, percent-encoding values.
 pub fn encode_query(params: &[(&str, String)]) -> String {
     let mut out = String::new();
     for (key, value) in params {
@@ -76,7 +56,6 @@ pub fn encode_query(params: &[(&str, String)]) -> String {
     out
 }
 
-/// Minimal RFC-3986 percent-encoding of query values.
 fn percent_encode(value: &str) -> String {
     let mut out = String::with_capacity(value.len());
     for byte in value.bytes() {
@@ -93,7 +72,6 @@ fn percent_encode(value: &str) -> String {
     out
 }
 
-/// Extracts the `data` field from a PBS `{ "data": ... }` response envelope.
 pub fn unwrap_data(body: &[u8]) -> Result<serde_json::Value, PbsError> {
     if body.is_empty() {
         return Ok(serde_json::Value::Null);
@@ -106,17 +84,12 @@ pub fn unwrap_data(body: &[u8]) -> Result<serde_json::Value, PbsError> {
         .unwrap_or(serde_json::Value::Null))
 }
 
-/// An established HTTP/2 PBS protocol session.
 pub struct H2Transport {
     send: h2::client::SendRequest<Bytes>,
     authority: String,
 }
 
 impl H2Transport {
-    /// Opens an upgraded h2 session at `/api2/json/{endpoint}?{session_query}`.
-    ///
-    /// `protocol` is the `Upgrade` protocol name (`proxmox-backup-protocol-v1`
-    /// for the writer, `proxmox-backup-reader-protocol-v1` for the reader).
     pub async fn connect(
         config: &PbsConfig,
         protocol: &str,
@@ -210,7 +183,6 @@ impl H2Transport {
         builder.body(()).map_err(transport)
     }
 
-    /// Awaits a response and collects its body, erroring on non-2xx status.
     async fn read_body(&self, response: h2::client::ResponseFuture) -> Result<Vec<u8>, PbsError> {
         let response = response.await.map_err(transport)?;
         let status = response.status();
@@ -244,7 +216,6 @@ impl H2Transport {
         unwrap_data(&self.read_body(response).await?)
     }
 
-    /// GET a path returning raw bytes (PBS `download`/`chunk` are not enveloped).
     pub async fn download(
         &mut self,
         path: &str,
@@ -255,7 +226,6 @@ impl H2Transport {
         self.read_body(response).await
     }
 
-    /// POST with query params and no body.
     pub async fn post(
         &mut self,
         path: &str,
@@ -266,7 +236,6 @@ impl H2Transport {
         self.read_response(response).await
     }
 
-    /// Sends a request carrying a binary body (e.g. a DataBlob), with flow control.
     pub async fn upload(
         &mut self,
         method: hyper::Method,
@@ -282,7 +251,6 @@ impl H2Transport {
         self.read_response(response).await
     }
 
-    /// PUT/POST with a JSON body (e.g. the dynamic-index digest/offset lists).
     pub async fn send_json(
         &mut self,
         method: hyper::Method,
@@ -297,7 +265,6 @@ impl H2Transport {
     }
 }
 
-/// Writes `data` to an h2 stream, respecting the peer's flow-control window.
 async fn send_with_flow_control(
     stream: &mut h2::SendStream<Bytes>,
     mut data: Bytes,
@@ -318,40 +285,4 @@ async fn send_with_flow_control(
 
     stream.send_data(Bytes::new(), true).map_err(transport)?;
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn host_port_parsing() {
-        assert_eq!(
-            parse_host_port("https://10.0.20.148:8007").expect("parses"),
-            ("10.0.20.148".to_string(), 8007)
-        );
-        assert_eq!(
-            parse_host_port("https://pbs.example.com").expect("parses"),
-            ("pbs.example.com".to_string(), 8007)
-        );
-        assert!(parse_host_port("ftp://x").is_err());
-    }
-
-    #[test]
-    fn query_encoding_escapes_reserved() {
-        let query = encode_query(&[
-            ("archive-name", "backup.tar.didx".to_string()),
-            ("ns", "a/b".to_string()),
-        ]);
-        assert_eq!(query, "archive-name=backup.tar.didx&ns=a%2Fb");
-    }
-
-    #[test]
-    fn unwrap_data_extracts_envelope() {
-        assert_eq!(
-            unwrap_data(br#"{"data": 42}"#).expect("parses"),
-            serde_json::json!(42)
-        );
-        assert_eq!(unwrap_data(b"").expect("empty"), serde_json::Value::Null);
-    }
 }
