+++ pbs-client/src/lib.rs
@@ -1,0 +1,12 @@
+pub mod auth;
+pub mod chunker;
+pub mod config;
+pub mod datablob;
+pub mod error;
+pub mod h2;
+pub mod manifest;
+pub mod naming;
+pub mod reader;
+pub mod rest;
+pub mod tls;
+pub mod writer;
