+++ pbs-client/src/manifest.rs
@@ -1,13 +1,5 @@
-//! The PBS backup manifest (`index.json.blob`).
-//!
-//! Field names are kebab-case to match PBS's `BackupManifest`/`FileInfo`. The
-//! manifest lists every archive/blob in the snapshot (but not itself), each with
-//! its size and checksum. We never sign or encrypt, so `crypt-mode` is always
-//! `none` and no signature is emitted.
-
 use serde::Serialize;
 
-/// File name of the manifest blob within a snapshot.
 pub const MANIFEST_BLOB_NAME: &str = "index.json.blob";
 
 #[derive(Serialize, Clone, Copy)]
@@ -22,7 +14,6 @@ pub struct FileInfo {
     pub filename: String,
     pub crypt_mode: CryptMode,
     pub size: u64,
-    /// Lowercase hex of the file's 32-byte checksum.
     pub csum: String,
 }
 
@@ -66,30 +57,7 @@ impl BackupManifest {
         self.files.push(file);
     }
 
-    /// Serializes the manifest to the JSON bytes that get wrapped in a DataBlob.
     pub fn to_json_bytes(&self) -> Result<Vec<u8>, serde_json::Error> {
         serde_json::to_vec(self)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn manifest_uses_kebab_case_and_hex_csum() {
-        let mut manifest = BackupManifest::new("host", "calagopus-test", 1_700_000_000);
-        manifest.add_file(FileInfo::new("backup.tar.didx", 1234, &[0xab; 32]));
-
-        let json =
-            String::from_utf8(manifest.to_json_bytes().expect("serializes")).expect("valid utf8");
-
-        assert!(json.contains("\"backup-type\":\"host\""));
-        assert!(json.contains("\"backup-id\":\"calagopus-test\""));
-        assert!(json.contains("\"backup-time\":1700000000"));
-        assert!(json.contains("\"crypt-mode\":\"none\""));
-        assert!(json.contains("\"filename\":\"backup.tar.didx\""));
-        assert!(json.contains(&"ab".repeat(32)));
-        assert!(json.contains("\"unprotected\":{}"));
     }
 }
